Refactor student/tutor letter strings as variables in CommandTestUtil class
There are several methods that make repeated use of "s", " s" and " s "
strings for students (and similarly, for tutors). The level of
abstraction in those methods for other strings is one level higher than
for the student/tutor letter strings. This is weak SLAP.

Abstracting out the student/tutor strings improves SLAP and makes the
code more maintainable and readable.

To do that, let's extract out the strings into a variable in
CommandTestUtil like the other strings for better consistency.